Crate solana_ledger

Modules

ancestor_iterator

bank_forks_utils

bigtable_delete

bigtable_upload

bigtable_upload_service

block_error

blockstore

The blockstore module provides functions for parallel verification of the Proof of History ledger as well as iterative read, append write, and random access read to a persistent file-based ledger.

blockstore_db

blockstore_meta

blockstore_metrics

blockstore_options

blockstore_processor

builtins

genesis_utils

leader_schedule

leader_schedule_cache

leader_schedule_utils

next_slots_iterator

rooted_slot_iterator

shred

The shred module defines data structures and methods to pull MTU sized data frames from the network. There are two types of shreds: data and coding. Data shreds contain entry information while coding shreds provide redundancy to protect against dropped network packets (erasures).

shredder 🔒

sigverify_shreds

slot_stats

staking_utils 🔒

token_balances

Macros

create_new_tmp_ledger

create_new_tmp_ledger_auto_delete

create_new_tmp_ledger_fifo

create_new_tmp_ledger_fifo_auto_delete

get_tmp_ledger_path

get_tmp_ledger_path_auto_delete

tmp_ledger_name