Module solana_ledger::blockstore_db

Modules

columns

Structs

Database

LedgerColumn

OldestSlot 🔒

PurgedSlotFilter 🔒

PurgedSlotFilterFactory 🔒

Rocks 🔒

WriteBatch

Enums

BlockstoreError

IteratorDirection

IteratorMode

Constants

ADDRESS_SIGNATURES_CF 🔒

Column family for Address Signatures

BANK_HASH_CF 🔒

Column family for bank hashes

BLOCKSTORE_METRICS_ERROR 🔒

BLOCKTIME_CF 🔒

Column family for Blocktime

BLOCK_HEIGHT_CF 🔒

Column family for BlockHeight

CODE_SHRED_CF 🔒

Column family for Code Shreds

DATA_SHRED_CF

Column family for Data Shreds

DEAD_SLOTS_CF 🔒

DUPLICATE_SLOTS_CF 🔒

ERASURE_META_CF 🔒

FIFO_WRITE_BUFFER_SIZE 🔒

INDEX_CF 🔒

Column family for indexes

MAX_WRITE_BUFFER_SIZE 🔒

META_CF 🔒

OPTIMISTIC_SLOTS_CF 🔒

Column family for optimistic slots

ORPHANS_CF 🔒

PERF_SAMPLES_CF 🔒

Column family for Performance Samples

PROGRAM_COSTS_CF 🔒

Column family for ProgramCosts

REWARDS_CF 🔒

Column family for Rewards

ROOT_CF 🔒

TRANSACTION_MEMOS_CF 🔒

Column family for TransactionMemos

TRANSACTION_STATUS_CF 🔒

Column family for Transaction Status

TRANSACTION_STATUS_INDEX_CF 🔒

Column family for the Transaction Status Index. This column family is used for tracking the active primary index for columns that for query performance reasons should not be indexed by Slot.

Traits

Column

ColumnName

ProtobufColumn

SlotColumn

SlotColumn is a trait for slot-based column families. Its index is essentially Slot (or more generally speaking, has a 1:1 mapping to Slot).

TypedColumn

Functions

get_cf_options 🔒

get_cf_options_fifo 🔒

Returns the RocksDB Column Family Options which use FIFO Compaction.

get_db_options 🔒

new_cf_descriptor 🔒

new_cf_descriptor_fifo 🔒

new_cf_descriptor_pair_shreds 🔒

Creates and returns the column family descriptors for both data shreds and coding shreds column families.

process_cf_options_advanced 🔒

should_disable_auto_compactions 🔒

should_enable_compression 🔒

should_exclude_from_compaction 🔒

Type Definitions

Result