Trait solana_ledger::blockstore_db::SlotColumn

pub trait SlotColumn<Index = u64> { }

SlotColumn is a trait for slot-based column families. Its index is essentially Slot (or more generally speaking, has a 1:1 mapping to Slot).

The clean-up of any LedgerColumn that implements SlotColumn is managed by LedgerCleanupService, which will periodically deprecate and purge oldest entries that are older than the latest root in order to maintain the configured –limit-ledger-size under the validator argument.