Struct solana_ledger::blockstore_meta::SlotMeta

pub struct SlotMeta {
    pub slot: Slot,
    pub consumed: u64,
    pub received: u64,
    pub first_shred_timestamp: u64,
    pub last_index: Option<u64>,
    pub parent_slot: Option<Slot>,
    pub next_slots: Vec<Slot>,
    pub connected_flags: ConnectedFlags,
    pub completed_data_indexes: BTreeSet<u32>,
}

The Meta column family

Fields
slot: Slot

The number of slots above the root (the genesis block). The first slot has slot 0.

consumed: u64

The total number of consecutive shreds starting from index 0 we have received for this slot. At the same time, it is also an index of the first missing shred for this slot, while the slot is incomplete.

received: u64

The index plus one of the highest shred received for this slot. Useful for checking if the slot has received any shreds yet, and to calculate the range where there is one or more holes: (consumed..received).

first_shred_timestamp: u64

The timestamp of the first time a shred was added for this slot

last_index: Option<u64>

The index of the shred that is flagged as the last shred for this slot. None until the shred with LAST_SHRED_IN_SLOT flag is received.

parent_slot: Option<Slot>

The slot height of the block this one derives from. The parent slot of the head of a detached chain of slots is None.

next_slots: Vec<Slot>

The list of slots, each of which contains a block that derives from this one.

connected_flags: ConnectedFlags

Connected status flags of this slot

completed_data_indexes: BTreeSet<u32>

Shreds indices which are marked data complete. That is, those that have the ShredFlags::DATA_COMPLETE_SHRED set.

Implementations

impl SlotMeta

pub fn is_full(&self) -> bool

pub fn is_connected(&self) -> bool

Returns a boolean indicating whether the meta is connected.

pub fn set_connected(&mut self)

Mark the meta as connected.

pub fn is_parent_connected(&self) -> bool

Returns a boolean indicating whether the meta’s parent is connected.

pub fn set_parent_connected(&mut self) -> bool

Mark the meta’s parent as connected. If the meta is also full, the meta is now connected as well. Return a boolean indicating whether the meta becamed connected from this call.

pub fn unset_parent(&mut self)

Dangerous. Currently only needed for a local-cluster test

pub fn clear_unconfirmed_slot(&mut self)

pub(crate) fn new(slot: Slot, parent_slot: Option<Slot>) -> Self

pub(crate) fn new_orphan(slot: Slot) -> Self