Struct solana_ledger::blockstore_metrics::BlockstoreRocksDbColumnFamilyMetrics

pub struct BlockstoreRocksDbColumnFamilyMetrics {

    pub total_sst_files_size: i64,
    pub size_all_mem_tables: i64,
    pub num_snapshots: i64,
    pub oldest_snapshot_time: i64,
    pub actual_delayed_write_rate: i64,
    pub is_write_stopped: i64,
    pub block_cache_capacity: i64,
    pub block_cache_usage: i64,
    pub block_cache_pinned_usage: i64,
    pub estimate_table_readers_mem: i64,
    pub mem_table_flush_pending: i64,
    pub compaction_pending: i64,
    pub num_running_compactions: i64,
    pub num_running_flushes: i64,
    pub estimate_oldest_key_time: i64,
    pub background_errors: i64,

}

A metrics struct that exposes RocksDB’s column family properties.

Here we only expose a subset of all the internal properties which are relevant to the ledger store performance.

The list of completed RocksDB internal properties can be found here.

Fields
total_sst_files_size: i64size_all_mem_tables: i64num_snapshots: i64oldest_snapshot_time: i64actual_delayed_write_rate: i64is_write_stopped: i64block_cache_capacity: i64block_cache_usage: i64block_cache_pinned_usage: i64estimate_table_readers_mem: i64mem_table_flush_pending: i64compaction_pending: i64num_running_compactions: i64num_running_flushes: i64estimate_oldest_key_time: i64background_errors: i64
Implementations

impl BlockstoreRocksDbColumnFamilyMetrics

pub fn report_metrics( &self, cf_name: &'static str, column_options: &LedgerColumnOptions )

Report metrics with the specified metric name and column family tag. The metric name and the column family tag is embedded in the parameter metric_name_and_cf_tag with the following format.

For example, “blockstore_rocksdb_cfs,cf_name=shred_data”.