Module solana_ledger::blockstore_processor

Structs

BatchExecutionTiming

Measures times related to transaction execution in a slot.

ConfirmationProgress

ConfirmationTiming

Measures different parts of the slot confirmation processing pipeline.

ExecuteBatchesInternalMetrics

PAR_THREAD_POOL 🔒

ProcessOptions

ReplayEntry 🔒

TransactionBatchWithIndexes

TransactionStatusBatch

TransactionStatusSender

Enums

BlockstoreProcessorError

TransactionStatusMessage

Functions

__stability 🔒

__static_ref_initialize 🔒

cache_block_meta

confirm_full_slot 🔒

confirm_slot

confirm_slot_entries 🔒

execute_batch

execute_batches_internal 🔒

fill_blockstore_slot_with_ticks

first_err 🔒

get_first_error 🔒

load_frozen_forks 🔒

process_bank_0 🔒

process_batches 🔒

process_blockstore_for_bank_0 🔒

process_blockstore_from_root

Process blockstore from a known root bank

process_entries 🔒

process_entries_for_tests

Process an ordered list of entries in parallel

process_next_slots 🔒

process_single_slot 🔒

rebatch_and_execute_batches 🔒

rebatch_transactions 🔒

run_final_hash_calc 🔒

schedule_batches_for_execution 🔒

supermajority_root 🔒

supermajority_root_from_vote_accounts 🔒

test_process_blockstore

verify_ticks 🔒

Verify that a segment of entries has the correct number of ticks and hashes

Type Definitions

CacheBlockMetaSender

ProcessCallback

Callback for accessing bank state while processing the blockstore