Struct solana_ledger::blockstore_processor::ProcessOptions

pub struct ProcessOptions {

    pub run_verification: bool,
    pub full_leader_cache: bool,
    pub halt_at_slot: Option<Slot>,
    pub new_hard_forks: Option<Vec<Slot>>,
    pub debug_keys: Option<Arc<HashSet<Pubkey>>>,
    pub account_indexes: AccountSecondaryIndexes,
    pub limit_load_slot_count_from_snapshot: Option<usize>,
    pub allow_dead_slots: bool,
    pub accounts_db_test_hash_calculation: bool,
    pub accounts_db_skip_shrink: bool,
    pub accounts_db_config: Option<AccountsDbConfig>,
    pub verify_index: bool,
    pub shrink_ratio: AccountShrinkThreshold,
    pub runtime_config: RuntimeConfig,
    pub on_halt_store_hash_raw_data_for_debug: bool,
    pub run_final_accounts_hash_calc: bool,

}

Fields
run_verification: bool

Run PoH, transaction signature and other transaction verifications on the entries.

full_leader_cache: boolhalt_at_slot: Option<Slot>new_hard_forks: Option<Vec<Slot>>debug_keys: Option<Arc<HashSet<Pubkey>>>account_indexes: AccountSecondaryIndexeslimit_load_slot_count_from_snapshot: Option<usize>allow_dead_slots: boolaccounts_db_test_hash_calculation: boolaccounts_db_skip_shrink: boolaccounts_db_config: Option<AccountsDbConfig>verify_index: boolshrink_ratio: AccountShrinkThresholdruntime_config: RuntimeConfigon_halt_store_hash_raw_data_for_debug: boolrun_final_accounts_hash_calc: bool

true if after processing the contents of the blockstore at startup, we should run an accounts hash calc This is useful for debugging.